fix unused import

+++ fastestimator/backend/_binary_crossentropy.py
@@ -91,7 +91,7 @@ def binary_crossentropy(y_pred: Tensor,
             ce = torch.nn.BCELoss(reduction="none")(input=y_pred, target=y_true.view(y_pred.size()))
 
         if class_weights is not None:
-            sample_weights = torch.ones_like(y_true, dtype=y_pred.dtype)
+            sample_weights = torch.ones_like(y_true)
             for key in class_weights.keys():
                 sample_weights[y_true == key] = class_weights[key]
             ce = ce * sample_weights.reshape(ce.shape)
